# Avoid pointers in SpanConfig, TracerConfig, EventConfig