Fixes #2065.

This does not appear to change any of the benchmarks in `sdk/trace`, however I would recommend this PR anyway because it feels more idiomatic to return structs, not pointers, for static configuration.

Before (at HEAD):
```
BenchmarkStartEndSpan/AlwaysSample-12    	 1348563	       876.4 ns/op	     848 B/op	       9 allocs/op
BenchmarkStartEndSpan/NeverSample-12     	 1633045	       734.8 ns/op	     752 B/op	       8 allocs/op
BenchmarkSpanWithAttributes_4/AlwaysSample-12         	  827178	      1465 ns/op	    1584 B/op	      17 allocs/op
BenchmarkSpanWithAttributes_4/NeverSample-12          	 1403071	       858.6 ns/op	     944 B/op	       9 allocs/op
BenchmarkSpanWithAttributes_8/AlwaysSample-12         	  599956	      1973 ns/op	    2112 B/op	      23 allocs/op
BenchmarkSpanWithAttributes_8/NeverSample-12          	 1271044	       938.9 ns/op	    1136 B/op	       9 allocs/op
BenchmarkSpanWithAttributes_all/AlwaysSample-12       	  643704	      1839 ns/op	    1936 B/op	      21 allocs/op
BenchmarkSpanWithAttributes_all/NeverSample-12        	 1319778	       907.7 ns/op	    1072 B/op	       9 allocs/op
BenchmarkSpanWithAttributes_all_2x/AlwaysSample-12    	  377712	      3129 ns/op	    3236 B/op	      32 allocs/op
BenchmarkSpanWithAttributes_all_2x/NeverSample-12     	 1000000	      1069 ns/op	    1392 B/op	       9 allocs/op
```

AFTER (this PR):
```
BenchmarkAttributesMapToKeyValue-12      	  452428	      2629 ns/op
BenchmarkStartEndSpan/AlwaysSample-12    	 1486172	       806.1 ns/op	     656 B/op	       7 allocs/op
BenchmarkStartEndSpan/NeverSample-12     	 1716672	       695.6 ns/op	     656 B/op	       7 allocs/op
BenchmarkSpanWithAttributes_4/AlwaysSample-12         	  857193	      1383 ns/op	    1392 B/op	      15 allocs/op
BenchmarkSpanWithAttributes_4/NeverSample-12          	 1473076	       812.7 ns/op	     848 B/op	       8 allocs/op
BenchmarkSpanWithAttributes_8/AlwaysSample-12         	  634714	      1882 ns/op	    1920 B/op	      21 allocs/op
BenchmarkSpanWithAttributes_8/NeverSample-12          	 1330555	       900.2 ns/op	    1040 B/op	       8 allocs/op
BenchmarkSpanWithAttributes_all/AlwaysSample-12       	  697209	      1723 ns/op	    1744 B/op	      19 allocs/op
BenchmarkSpanWithAttributes_all/NeverSample-12        	 1379874	       886.3 ns/op	     976 B/op	       8 allocs/op
BenchmarkSpanWithAttributes_all_2x/AlwaysSample-12    	  404030	      3066 ns/op	    3044 B/op	      30 allocs/op
BenchmarkSpanWithAttributes_all_2x/NeverSample-12     	 1000000	      1001 ns/op	    1296 B/op	       8 allocs/op
BenchmarkTraceID_DotString-12                         	17515768	        68.86 ns/op
BenchmarkSpanID_DotString-12                          	22704256	        53.28 ns/op
```