@MrAlias 2021-08-23
It would be helpful if we can show this change will reduce the number of allocation in a benchmark. The existing benchmarks show no change, but I suspect this is because they are not testing what is being changed.

@jmacd is this something you can do or would you like some help with this?