@jmacd 2021-08-23
I'd like to help. To take this further, I ran 

```
go test -bench=BenchmarkStartEndSpan/AlwaysSample -test.benchtime=10s -test.memprofile=mem.out
```

and looked at the output with

```
go tool pprof -alloc_objects -lines ./mem.out
```

The result after my previous PR shows 9 allocations, paraphrased as:

1. the new Context.Value
2. the start `SpanConfig` escapes
3. the end `SpanConfig` escapes
4. the new `*span` 
5. crap load of attribute map stuff (5 allocs)

I was able to modify this PR to produce the benchmark improvement. I'd like help completing this PR from someone in the group, if possible. The PR is mostly complete, but the same fix should be done in a bunch of places for consistency. `TracerConfig` and `EventConfig`.

The fix is a little ugly, too. (I've done this before, 🤷.)

The option functions can't take pointers, so you end up with Options like:

```
func (o timestampOption) applySpan(c SpanConfig) SpanConfig {
	c.timestamp = time.Time(o)
	return c
}
```