@jmacd 2021-08-23
Also, if we're looking for opportunities to improve allocations, the crap load of attribute map allocs are worth looking at.

![image](https://user-images.githubusercontent.com/3629705/130521869-16b46e59-9f2c-4551-bdaa-64df6c890729.png)