@jmacd 2021-08-23
You can see the 7 resulting allocs in the picture above. (5 of them are the attribute map/list!)