@Aneurysm9 2021-08-24
> The option functions can't take pointers, so you end up with Options like:
> 
> ```
> func (o timestampOption) applySpan(c SpanConfig) SpanConfig {
> 	c.timestamp = time.Time(o)
> 	return c
> }
> ```

I'm a bit confused by this.  How are the option functions not able to accept pointers and how does taking the config struct by value and returning a modified copy result in fewer allocations?