@MrAlias 2021-08-30
Superseded with https://github.com/open-telemetry/opentelemetry-go/pull/2212

The overall memory allocation issue will be addressed in a follow on to #2212 and will likely be based on the work found in this PR that will address #2065.